requirements.txt
anticaptchaofficial==1.0.59
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chrome-cookiejar==0.1.0
chromedriver-autoinstaller==0.6.2
cryptography==42.0.7
exceptiongroup==1.1.3
git-filter-repo==2.38.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
gspread-dataframe==3.3.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
iso8601==2.1.0
kaitaistruct==0.10
m3u8==3.6.0
m3u8-To-MP4==0.1.11
numpy
oauthlib==3.2.2
outcome==1.2.0
packaging==24.0
pandas
py==1.11.0
pyasn1==0.5.1
pyasn1_modules==0.4.0
pycparser==2.21
pycryptodome==3.19.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyVirtualDisplay==3.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.11.2
selenium-stealth==1.0.6
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.8.0
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.0.4
webdriver-manager==4.0.1
websockets==12.0
whitenoise==6.6.0
wsproto==1.2.0
zstandard==0.22.0

